Move NowcastSource into data pipeline
