+++ pipelines/data/nowcast.py
@@ -1,7 +1,13 @@
-"""Shared nowcast data objects."""
+"""Shared nowcast data objects and source protocol."""
+
+from __future__ import annotations
 
 import datetime as dt
 from dataclasses import dataclass, field
+from typing import TYPE_CHECKING, Protocol
+
+if TYPE_CHECKING:
+    from pipelines.epiautogp.forecast_spec import ForecastSpec
 
 
 @dataclass(frozen=True)
@@ -10,3 +16,24 @@ class NowcastData:
 
     dates: list[dt.date] = field(default_factory=list)
     reports: list[list[float]] = field(default_factory=list)
+
+
+class NowcastSource(Protocol):
+    """Interface for producing nowcast data for a forecast configuration."""
+
+    def applies_to(
+        self,
+        *,
+        forecast_spec: ForecastSpec,
+    ) -> bool:
+        """Whether this source applies to a forecast configuration."""
+        ...
+
+    def get_nowcast_data(
+        self,
+        *,
+        dates: list[dt.date],
+        reports: list[float],
+    ) -> NowcastData:
+        """Estimate nowcast data for one model run."""
+        ...
